Time: 27 ms (98.40%), Space: 5.9 MB (82.89%) - LeetHub

+++ 1680-concatenation-of-consecutive-binary-numbers/1680-concatenation-of-consecutive-binary-numbers.cpp
@@ -4,7 +4,7 @@ public:
         long long int MOD=1e9+7;
         long long int ans=0;
         long long int shift=0;
-        for(long long int i=1;i<=n;i++) {
+        for(int i=1;i<=n;i++) {
             if((i&(i-1))==0) {
                 shift+=1;
             }
